Fix: 1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -9,7 +9,7 @@
 unsigned long int _pow(int power)
 {
 	int i;
-	unsigned long int num = 1;
+	unsigned long long int num = 1;
 
 	if (power == 0)
 		return (1);
